Player: Delete decoder after decoder thread stopped
It is important to delete the decoder in DecoderEnd to ensure that any
threads allocated by the Ffmpeg decoder are stopped before the video
buffers are deallocated.  If not then random crashes can be observed
when exiting playback of H264 encoded material if CPUs are > 1.

Fixes #11452

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>
Signed-off-by: Stuart Morgan <smorgan@mythtv.org>
(cherry picked from commit 6e62a3c601b93c6d9525cc503634350a965a5058)