Have mythfilldatabase open a blocking connection to master backend
This opens a blocking connection that lasts the entire duration
mythfilldatabase is running, preventing the backend from deciding to
shut down during that time. It will continue to run even if the initial
connection fails.

There is existing behavior in mythshutdown to test for a running
mythfilldatabase, but this seems something that should be internal.