MythMusic: Attempt to keep playing on error
Previously if the music player encountered an error it would stop playing. Now
it will try to continue playing the next track in the playlist. If it fails to
play 5 tracks in a row then it will still stop though since it's likely there
is some serious problem like an NFS server can't be reached.

This also cleans up the error reporting and changes it to use the new
notification popups.

Fixes #11681.