mythweb does not handle MESSAGEs from the backend
This changes the announce to not receive any events at all.
I could not find a reference to any uses of events with PHP, if
there are some we can turn this into a parameter.
This fixes the recording list on mythweb erroring out with the
option to select the series "UPDATE_FILE_SIZE 4092 2009-09-29T02:58:42 290586314"
if a recording is in progress and the event comes down the socket before
the list of recorded programs.

Refs #11290