UI: Fix Xv painter rendering of OSD menu gradient elements
Commit #xxxxx re-worked and simplified MythPainter.
MythYUVAPainter::DrawRect is passed a brush to paint the background.
This brush may be configured to gradient fill - for instance in
the MythCentre theme to draw the MythPopupBox OSD menu buttons.
Currently only the basic brush colour is translated to YUVA so the
OSD menu is incorrectly rendered.

This patch translates the colours in the brush fill gradient from
RGB to YUV.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>
Signed-off-by: Stuart Morgan <smorgan@mythtv.org>