Don't wait for paused audio to drain.  Fixes #11730.
The problem is that MythPlayer::AVSync() might pause audio while
dropping a frame, and if this happens during the EOF A/V draining
period, it could wait forever for the audio to drain.

Alternatively or in addition, AVSync() could check that
GetEof()!=kEofStateDelayed before pausing audio.  Thanks to Mark
Spieth for the analysis.