Make signal haveResult(metadata) thread-safe.
Fixes potential leaks should there was no listener set. If there was more than one listener at a time it would have crashed
This change the definition syntax for RefCountHandler. For some reasons Qt wants a default constructor to be defined so an object can be defined as a qMetaType.
It's actually much better that way, allows for multiple-listeners and prevent to send pointers across thread/classes which is a bad idea to start with

Ref #11754