Disable database logging by default.
It can be re-enabled using --enable-dblog

Until we actually start using this data somewhere there is not much
point in storing it. Since it's already known that writing to the
database can affect I/O performance, especially where drives are
shared with recordings, this could potentially even cause a positive
feedback loop. The I/O starvation causes the backend to start logging
slow read/write warnings, which in turn increase the writes to the
database exacerbating the problems.

The same problem is also true, to a greater or lesser extent with file
based logging but we actually need the file logs so there isn't much
we can do there.