Make some Playback OSD dialog timouts more sane.
Changes the OSD dialog timeouts to Infinite for:
  Adjust Audio Sync
  Adjust Subtitle Delay
  Adjust Subtitle Zoom
  Adjust Time Stretch

As many users have observed, for all of these dialogs, it can take a
while of observation to determine whether the adjustment is good
enough, and it's annoying if the dialog suddenly disappears in the
meantime, especially if you then press an error key and inadvertently
jump to a different position.

In addition, for all of these dialogs, the SELECT key now closes the
dialog.  (This was already true for Adjust Time Stretch.)