MHEG: Avoid an Interaction channel deadlock if data is cached
If a request to NetStream is pending when a 2nd is requested that can be
resolved from the cache then a deadlock occurs if the 1st request
finishes while the cached request is processed.

This patch releases the MHInteractionChannel mutex while the
synchronous request for cached data is processed.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>
Signed-off-by: Stuart Morgan <smorgan@mythtv.org>
(cherry picked from commit 30f1e586197a60ee58f0711dca0fd3c3b18e3e4a)