MythMusic: Filter cached cddb discIDs to prevent incorrect CD identification
In cddb.cpp, Dbase::CacheGet uses the QMap::find function to iterate
over the cached cddb records that match a given discID.  The Qt docs
for 4.7 say that incrementing the iterator returns the next QMap value
with the requested key but in testing values with different keys can be
returned.

This patch works around this Qt bug by verifying the key.

Fixes #10892

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>
Signed-off-by: Stuart Morgan <smorgan@mythtv.org>