Speed up drawing caption text with outlines.
The old method of drawing text with outlines involves drawing the same
text many times at various offsets, with the offsets based on
traversing the perimeter square of pixels with edge length 2N+1, where
N is the specified outlinesize.  This means drawing 8N extra copies of
the text.  With some fonts and point sizes, this can be a very
expensive operation, leading to video stuttering while captions are
drawn.

The new method draws just one extra copy, but with a pen width set to
2N+1, which is much faster.  Beware that using a huge outlinesize may
lead to strange artifacts in the outline (but that was already the
case for the original outline code).

Refs #10984.