Make sure smolt directory is created before trying to read from it.
This makes sure the configuration storage directory has been created
prior to attempting to use it. Smolt typically stores to /etc, and
assumes the packager has created its desired directory. There is
probably a cleaner place to put this code, but this is a catch-all for
several files read from that directory.

Fixes #11803