Improve performance of processing CEA-708 captions.  Refs #10984.
Change the complexity of creating cc708 strings from quadratic to
linear.

Allow leading and trailing strings of spaces to be joined to their
neighbor strings even when the pen is different, to reduce the number
of chunks to process.  (These chunks would have been eventually
optimized away before drawing, but now the overall processing time is
reduced.)