MythArchive: Catch StopIteration exception in mythburn.py
If a DVD is created from a single mpeg video then mythburn.py attempts
to transcode the file to eliminate any errors.  However, runMythtranscode
calls DB.searchRecorded, to obtain any cutlist, with an invalid chanid
and starttime which results in a StopIteration exception.

This patch catches the exception and returns an empty cutlist.

Fixes #10758.

Signed-off-by: Paul Harrison <pharrison@mythtv.org>