Revert "MHEG: Fixes for InteractionChannel streaming from network URI's"
This broke playback of all videos for a number of people. It seems
like there was probably a deadlock somewhere in the mhi code on
startup. Reverting the whole thing since this has such a serious
impact, but it may be re-applied later when we've identified and fixed
the issue.

Fixes #11831

This reverts commit ea77f42462bd1055578131fb64e1586563db9442.