AFD: Read packets until audio PTS >= video PTS
If the 'Extra audio buffering' setting is disabled then the audio
stream PTS can lag the video by up to a second.  This can sometimes
cause A/V sync problems but also causes sync problems when displaying
the audio output graph when editing the cutlist.

This patch ensures that the last audio PTS at least equals the video
PTS before exiting GetFrame.  This ensures there are matching audio
samples for the pause frame shown in the cutlkist editor.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>
Signed-off-by: Jean-Yves Avenard <jyavenard@mythtv.org>