Have pager return Nones if the server returns null.
This temporarily resolves an issue where the TheMovieDb server returns
nulls in searches when a movie has been deleted but not yet flushed out
of the cache. This prevents the error internally, but does not attempt
to re-order to remove the nulls, so is merely passing the issue onto the
application using this library.

Refs #29