Have pager return Nones if the server returns null.
This temporarily resolves an issue where the TheMovieDb server returns
nulls in searches when a movie has been deleted but not yet flushed out
of the cache. This prevents the error internally, but does not attempt
to re-order to remove the nulls, so is merely passing the issue onto the
application using this library.

Refs #29
(cherry picked from commit 0d91e6fc4184c17c5a9d39e1b2a264602a47747d)