UPNP: Remove use of <URLBase> from Device Description xml.
As noted by Rune Petersen in #10696 the use of a URLBase element in
the device description is not recommend in the UPnP 1.0 spec. (UPnP
1.0 Device Architecture document, page 27). Additionally it's use is
forbidden by DLNA, although DLNA devices MUST be able to work with
servers using it.

This potentially fixes issues with Samsung Smart TVs, although their
behaviour is definitely a bug in their firmware and is it not UPnP or
DLNA compliant.

Fixes #10696