Don't send ALLOW/BLOCK_SHUTDOWN on the event socket. Refs #11777
The event socket is a 'Monitor' connection it won't block shutdown so
there's nothing gained by sending these messages on it. Additionally
it's only ever used in conjunction with a control socket to the same
server (master) and we've just sent the message via that socket, so
it's redundant.
(cherry picked from commit cb8b4204808879e1e7205bdf900e89e71f3e08a2)