Allow system calls to pull global settings from the database.
This adjusts the System() class to first try local settings of the
supplied value, fall back to global settings, and then the supplied
default, before finally erroring. This corrects an issue where the
VideoGrabber() child class did not properly pull the global settings for
TelevisionGrabber and MovieGrabber, and thus just used the internal
defaults.