MythMusic: fix playback of ogg radio streams
Peek the probe data from the stream buffer rather than read the data otherwise
ffmpeg fails to open the stream because the magic it is looking for has already
been read from the stream.

Also increase the probe buffer to 16K this seems to be required for some ogg
streams to be probed properly.