Change RI::ForgetHistory() to use dupmethod when available.
Previously, ForgetHistory() implicily used subtitle and description as
the duplicate check method.  Consequently, it doesn't always match
correctly when other methods are used and there are guide changes in
data that should be ignored.  In addition, don't assume any dupmethod
in the Previously Recorded screen to avoid being to overzealous when
called from there.