AFD: Fix playback of MPEG1 recordings with Xv and VAAAPI
1. Ffmpeg's avcodec_open2 fails if the stream's type doesn't match
the decoder's type.  This patch forces the codec type to mpeg2 as
well as the stream to work around this limitation.

2. Ffmpeg cannot decode MPEG1 to a VAAPI supported pixel format so it
is necessary for VAAPI to refuse to handle an MPEG1 stream.  In this
case the decoder will fallback to Xvideo.

Without this patch the playback of mpeg1 videos is broken.
This bug affects older videos and cameras using MPEG1 encoding.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>