Add mythutil --getmarkup and mythutil --setmarkup commands.
This extracts the contents of the recordedmarkup, recordedseek, and
filemarkup tables to or from a local xml file.  The primary reason is
to allow users to extract recorder seektable information as needed and
attach it to a ticket, for a developer to use for debugging.  Example:

  mythutil --chanid xxx --starttime yyy --getmarkup /tmp/markup.xml
  mythutil --video zzz.mpg --setmarkup /tmp/markup.xml

An added benefit is to be able to copy metadata (e.g. seektable,
commskiplist, cutlist) between isolated MythTV instances.
(cherry picked from commit dafb81abcf8955112037d2e7e230ebb8eb068886)