Add secondary check to prevent housekeeper task from running twice.
This adds an extra check to prevent a HouseKeeperTask from running
more than once simultaneously. This prevents an issue seen in a
misconfigured child task that was queued in concurrent passes. One
instance had not yet finished when the second one started a minute
later, and overwrote a shared pointer, eventually resulting in a
segmentation fault.

Cause spotted by Jonatan Lindblad.

Refs #11902