Add missing toUTC() conversions in the preview generator.
Without these conversions, modification timestamps on
backend-generated preview files are reported in local time, but
interpreted by the frontend as UTC.  The frontend then converts this
to local time and uses that to set the modification timestamp of the
locally cached preview, effectively back-dating the cached file in
most time zones.  This messes up date comparison logic for determining
whether to regenerate or reload the preview image.