Allow Live TV to be entered from the top-level Program Guide.
This allows the user to start Live TV on a specific channel from the
Program Guide grid.  Otherwise, the user needs to wait for Live TV to
start and tune the initial channel, then find the channel actually
wanted, and wait for it to be tuned.  This also provides a workaround
when the default initial channel is untunable for some reason.

Currently, if the desired channel is not tunable (e.g., all capable
recorders are already busy), it will fail.

Refs #11913.

(cherry picked from commit c0c5299ab82f638157b17ba13af154f73c27a40d)