Libmythbase: Stop MythWelcome blocking BE idle shutdown
MythWelcome connects to the backend server using a "Monitor"
announcement so as not to block backend idle shutdown.  However,
if the connection is broken for whatever reason e.g. if the
backend is stopped during STR, then when the connection is
re-established it currently uses "Playback" mode if the idle
shutdown time is > 0.  This causes the MythWelcome connection
to block further idle shutdowns.

This fix passes the original blocking mode to ConnectToMasterServer
rather than idleTimeoutSecs > 0

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>