Adds Server Side Import (include) for scripts.
This is for server side execution only.  If an import statement
is found at the start of a line, the files contents are injected
into the current server side script that's being processed.

No further processing is done on the imported file
(it must be valid javascript code).

Syntax:
   - must be at start of line (no leading spaces)
<%
import "/relative/path/to/script.js"

%>

 - or -

<%import "/relative/path/to/script.js"%>