Hook up the 'quick record' menu in the Web Guide
The quick record menu allows simple schedules to be created in just
two clicks without ever leaving the guide. First click on the program
you want to record, a menu will appear offering 4 different scheduling
options, "Record this showing", "Record all showings", "Record one
showing" and "Record all showings on this channel". A fifth button
will take you instead to the scheduler editor (ultimately).

There is still no schedule editor, and no option yet to cancel a
schedule/recording. The latter will be the next thing I add.

This involves a bit of a hack with a timer to check the result of the
reschedule. This will ultimately be replaced with a websocket based
solution but it works for now, assuming your MATCH reschedule requests
take less than 2.5 seconds.

Generally the code could use a cleanup, but it works fine as a
demonstrator.