Be more flexible about starting the Guide at a particular channel.
The DefaultTVChannel specifies the initial channel to highlight when
the Program Guide is started from the top level (not from within Live
TV).  The original logic required an exact match on the channel
number, and if there was no exact match, it would start at the
beginning of the channel list.

The new logic finds the closest match using the same "smart" channel
number comparison function used for sorting the Guide by channel
number.

Refs #11913.