VO: Prevent VAAPI scibbling in CropToDisplay
When using VAAPI decoding, VideoOutputOpenGLVAAPI::ProcessFrame is
called from MythPlayer::VideoLoop to display a frame.  This in turn
calls VideoOutputOpenGL::ProcessFrame which calls VideoOutput::CropToDisplay
However, this function only handles YV12 format frames and not VAAPI
format and can write beyond the end of the video buffer.

This patch checks the frame codec and returns early if the format
is not YV12.  This does not affect any other formats as OpenGL
normally uses YV12 format.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>