VO: Fail VAAPI reinit from decoder thread to avoid Qt abort in doneCurrent
If a recording starts without a video stream then if the video
subsequently becomes active the decoder attempts to re-init the
display.  However, the OpenGL functions fail if called outside of the
UI thread and this results in an unbalanced series of calls to
doneCurrebt which triggers a Qt abort (debug build) due to a recursive
mutex overflow.

This bug is tickled by the CloseShave video which started recording
at just before 7pm on BBC3, which is just before the start of normal
BBC prgrams.  The initial stream consists of MHEG data only.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>
Signed-off-by: Jean-Yves Avenard <jyavenard@mythtv.org>