Initial work on the Recordings page for the internal server
Please be aware that the first time you view the page it will create
previews for all the listed programs, this may take a very long
time and create a high system load. A fix for this will hopefully be
devised, although it should only be an issue the first time you view the page.