MythZoneMinder: fix playback of continuous events
Not every frame is stored in the database for continuous events so their
playback was broken. This fixes the problem by reading the number of frames
and the event time from the DB and interpolating the frame list and time deltas
from that.

Refs #8912

(cherry picked from commit 68f98331c2e2053e803d6237eacb75a7294c2825)