Allow the Guide start channel to be the "natural" Live TV channel.
If the DefaultTVChannel setting is left blank, then the EPG starting
channel is dynamically selected to be the same channel that would
naturally be tuned when Live TV is started.  This depends on the
available cards, their Live TV priority ordering, and the current
value of the starting channel for the inputs.

Note: It looks like there are still problems selecting the correct
tuner and starting channel when starting Live TV, which need to be
separately fixed.

Refs #11922.

(cherry picked from commit eb13dd5af48535c64196c9c6aa5e5235f37f6555)