Remove the option to start Live TV in the Program Guide.
The user can instead start Live TV and then bring up the Guide, or
start the Guide and then select "Watch This Channel" from the menu.

With the first option, to get to the Guide quickly, the user may want
to use the "Program Guide" jump point, or reorganize their Playback
OSD menu or Playback OSD compact menu
(http://www.mythtv.org/wiki/Themed_Menus) to bring the Program Guide
item to a more convenient position in the menu.

Refs #11922.