WebFrontend: Add Lazy Loading (jScroll) to Recordings and Upcoming Recordings page.
I started to add it into the guide, but due to both a services
limitation which needs fixing first and the added complexity of having
to remember state while browsing forward and backward in time through
the guide it's currently not functioning and is disabled.

Packagers: I had to fix a bug in jScroll before it would work correctly in our
environment. I've submitted a patch upstream. Please do NOT use a
package version of jScroll instead!