MythZoneMinder: by default exclude continuous events from 'Show Events' screen
You can now toggle whether continuous events (events from monitors that are
continuously recording) are shown in the list of events. Press Menu and select
the option to toggle the setting which will be remembered whenever you enter
the 'Show Events' screen.

If you select an individual monitor then all events will still be shown for the
selected monitor this only filters events when showing events for all monitors.

Note: bumps the protocol version so both mythzmserver and the plugin need to be
updated.