Services: Include conflicting recordings in the GetUpcomingList() output even when ShowAll isn't true.
I don't think this was the original intention, but if we are to mirror
the behaviour of the frontend, mythweb etc then this list should
include conflicts. This may yet evolve into a seperate ShowConflicts
argument so both the old and new behaviour are possible.

We already show aborted, failed and other 'bad' states in this list.