configure: Special case libfreetype test
Include the freetype header, in-directly through a macro, like it
is done in the drawtext filter. Do not break if the header is moved.

Unfortunately the drawtext filter included the file where the include
macros are defined in a wrong way. This is not needed and breaks the
build. Remove that #include line too.

(cherry picked from commit cea5812fa723c08b89d929eeba73462e05de2973)

Signed-off-by: Alexander Strasser <eclipse7@gmx.net>

Conflicts:
	configure
Signed-off-by: Jean-Yves Avenard <jyavenard@mythtv.org>