MetaIO: validate the files extension before creating a tagger
This checks the file extension is one we know how to handle before
creating the tagger. Return NULL if not valid.

This fixes a bug with the music file scanner trying to add files with an
unknown file extension.

NOTE: bumps the ABI version so compile and install myth core before compiling
the plugins