MythUtil: add an --updatemeta option to update music tags
This updates both the database record and the tracks tag with the
new metadata. A MUSIC_METADATA_CHANGED mythevent is broadcast so
all FE's will automatically update to relect the new track details.

See 'mythutil --help updatemeta' for more details.

Note: the music utils are a work in progress so subject to change.