MythMusic: start decoupling the file scanning from the frontend
This allows the scanning of music files to work outside of the FE. This is
necessary to switch MythMusic to use storage groups since the music may not be
stored on the same host as the FE.

The file scanner is now more fire and forget so the progress dialogs have been
removed. It now sends out events announcing the start and end of a scan
which also includes a summary of the scan results which any clients listening
can use to inform the user what is happening. eg. the FE will use notifications
to let the user know a file scan has started and when it completes another
notification will give a summary of the results of the scan.