RemoteFile: add a static FindFile() function to find a file on all BE's
This is a helper function that will search all BE's for a file in a
storage group. If a host is given the search will use that as a hint and
search that host first. If host is empty the search will look on the master
BE first. Returns the myth URL that can be used to access the file or an empty
string if not found.

Note: bumps the API version so both core and plugins need to be recompiled.