Add channel callsign to OSD
After adding a HDHR Prime to my system it became apparent that the OSD did not display the channel callsign which made browsing channels very difficult. This seemed to work well for me. The SD theme could probably use this too but the widget placement would need to be tweaked.