Audio: Offload dbase updates when volume is changed
VolumeBase::SetCurrentVolume writes the new volume immediately to the dbase.
This dbase update occurs on the main UI thread and can take several mS
on a remote FE.  The delay caused by a series of volume changes (e.g. by
holding down the volume up/down button) can adversely affect video playback,
causing minor loss of a/v sync which results in the player dropping frames
which in turn is distracting to the viewer.

This patch delegates volume dbase updates to a separate thread which
also adds a holdoff period in order to reduce the number of updates.

LVR 0.27/0234
Fixes #12144

Signed-off-by: Roger Siddons <rsiddons@mythtv.org>