make parameter names consistently lowercase
Since [3d9155909f] the parameter names used inconsistent upper and lower case.

There are two places in libmythupnp where the incoming packets are decoded, in
one place they are converted to lowercase, in the other place they are not.
This commit had made it lowercase in both places but did not change the code
that accessed the data, the caused serious discombobulation and the end of UPNP
watching through out the world.

Patch by Steve Beinicke

Fixes #12006