ImportMusic: update the music import to use storage groups
This allows the import music screen to move the imported tracks to any backend
with a directory in the 'Music' storage group.  The master will be used by
default but you can save to any BE by using the "Select Where To Save Tracks"
option in the popup menu. The option is sticky so once changed the same one
will be used the next time until it's changed.

Still need to work on the album art side of things.