Make "DailyArtworkUpdates" a global option
This makes the DailyArtworkUpdates setting a global setting so it can be
properly configured on any frontend. This patch does not include any
code to fix an existing improperly configured setting, which could cause
problems in the future, as the per-host option will override the global
one and will be uneditable without direct database modification.

Refs #12033