MythMusic: remove the local file access from findIcon()
This updates findIcon() to only look for the icons in the AlbumArt storage
group. Currently it only looks on the master I may change this to also look
on any slaves as well.

This also adds caching of the result of the search in a map to speed up the
lookups because searching for files over the myth protocol can be quite slow.