MythMusic: update the music cd ripper to use storage groups
The encoders and taggers require local file access so we now use a temp
directory in MythMusic's config directory to rip the tracks to and once
complete they are copied over to any backend with a directory in the
'Music' storage group.  The master will be used by default but you can save
to any BE by using the "Select Where To Save Tracks" option in the popup
menu. The option is sticky so once changed the same one will be used the next
time until it's changed.

Note: Due to  limitation in RemoteFile::DeleteFile() you cannot remove any
tracks on slave BE's from the rip screen so if you need to re-rip an existing
track on a slave BE you will have to manually delete it first.