Prevent playback to exit early should backend's writes be delayed
Typical scenario is when using liveTV and being very close to real time TV: should the backend take more than 360ms between two writes, the reader would treat it as an EOF and abort the read.

Whenever we open a file in write mode, we add it to a global dictionary. When reaching an EOF, we can then check if the file is going to be written to again.

We can also remove all EOF retries in FileRingBuffer::safe_read() as we definitely know now if we did reach EOF (as we check against the file size).
This allows for almost instantaneous program transition in live TV.