Allow seeking within the entire liveTV chain of recordings.
Seeking across liveTV was fundamentally broken. You could only seek within the current recordings or go to the beginning of the next one or 15s to the end of the previous one.
This changes correct this behaviour. It is now possible to seek across the entire live TV session. We ensure however that we do not seek further than 15s from the end of a recordings (3s if reaching end of live program).

Fixes #12076

(cherry picked from commit 11daeaadc52214605e204cdfe6f7c001c9c1cc9b)