Apply preroll on virtual tuners when the same multiplex is in use.
Previously, if any tuner in the same input group was already busy,
preroll would not be applied.  Now, if the recording on the other
input is on the same multiplex as the new recording, preroll will be
applied.

Refs #12023

(cherry picked from commit 11e65dfd38ca419b2157f0dbaa19f965647f648c)