MythMusic: add a 'First' resume mode and make 'Off' never resume
This fixes the resume playback feature so that 'Off' never starts playback and
adds a new 'First' mode which starts at the first track in the active playlist.

'Track' and 'Exact' modes remain the same and resume playback at the start of
the track or exact position within the track that was playing when MythMusic
was last exited.