Add the ability to scan for the programs in a MPTS stream from an IPTV source.
When scanning for channel, there is now a "M3U Import with MPTS" Scan Type
option.  If that is selected instead of "M3U Import", then after reading the
m3u to find the url, the stream will be parsed to find available programs,
which can then be inserted into the channel table.

To develop this, I installed the tools from OpenCaster
(http://www.avalpa.com/the-key-values/15-free-software/33-opencaster), and
used them to broadcast an MPTS file:

$ mkfifo fifo1.ts
$ mkfifo fifo2.ts
$ tsloop sample.mpts > fifo1.ts &
$ tsstamp fifo1.ts 95000000 > fifo2.ts &
$ tsudpsend fifo2.ts 127.0.0.1 5001 95000000

In myth, I created a "Freebox" UDP recorder with an m3u file which looks
like:

~~~~~~~~~~~~~~~~~~~~~~~~~
#EXTM3U
#EXTINF:0,5001 - UDP
#EXTMYTHTV:xmltvid=udp
udp://127.0.0.1:5001
~~~~~~~~~~~~~~~~~~~~~~~~~

I was then able to scan for the progams in the MPTS stream.

To create that sample.mpts file, I enable the "Record Full TS" option for
the "High Quality" profile, and then used that profile to record all the
subchannels available on one of my ATSC stations.  See [b0aa4e50c]