Decoder: Limit buffered audio data to 2000mS for visualiser startup
The decoder thread repetitively calls AvFormatDecoder::GetFrame to
decode incoming audio and video packets.  The decoder sleeps when video
or audio buffers are full.  With an audio only stream the decoder runs
until the audio buffer is full.  This can end up with 10-15 seconds of
buffered, decoded audio data.

When the visualisation is enabled, new frames are buffered to a max of
500 when the visualisation is disabled.  If there are more than 500
frames buffeed then the visualisation is never shown

This patch sets an upper limit of 1000msS of buffered audio data.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>
Signed-off-by: Jean-Yves Avenard <jyavenard@mythtv.org>